Module Eval

eval.ml

include Eval__.Vm

type vm_atom = string * link array

and link =

| NormalLink of int * vm_atom Stdlib.ref
(port_i, atom)

type register = vm_atom Stdlib.ref

val null_ptr : vm_atom Stdlib.ref

レジスタを初期化するためだけのアドレス（本来はいらない）

val null_link : link

リンクを初期化するためだけのアドレス（本来はいらない）

type registers = register array

レジスタ

val init_register : int -> registers

初期状態のレジスタを確保する

val get_normal_link : link -> int * vm_atom Stdlib.ref

NormalLink コンストラクタを剥がす

val free_atom : (string * 'a) Stdlib.ref -> unit

Free memory fragment of the given address. Possibly implemented with option type and assign None.

module AtomLists = Eval__.Vm.AtomLists

アトムリストの Map．

val dump_link : (vm_atom Stdlib.ref * 'a) list -> link -> int * 'a

リンクの解消

val dump_atom : (vm_atom Stdlib.ref * 'a) list -> 'b -> ('c * link array) -> 'b * ('c * (int * 'a) list)

val dump_atom_list : vm_atom Stdlib.ref list AtomLists.t -> (int * (string * (int * int) list)) list

val reduce : register list Eval__.Vm.AtomLists.t -> (string * (int * (Generator.lhs_inst list * Generator.rhs_inst list))) -> (string * register list Eval__.Vm.AtomLists.t) option

Try to reduce one step with the given atoms and a rule

ルール適用に成功したら Some で包んだ更新された atom_list を返す

val run_once : register list Eval__.Vm.AtomLists.t -> (string * (int * (Generator.lhs_inst list * Generator.rhs_inst list))) list -> (string * register list Eval__.Vm.AtomLists.t) option

Try reduce one step with the given atoms and rules

val init_atoms : (int * Generator.rhs_inst list) -> register list Eval__.Vm.AtomLists.t

push the initial graph and return their references